Use different levels in eco and overflow modes
In eco mode, the tank level must be higher because it must support going
to the overflow mode which will move a lot of water from the tank to the
pool. In overflow mode, it is normal to have the tank at a lower level
without need to refill it.