New eco mode
The original eco mode was not optimal. This is an improved version. It
splits the filtration over several periods during the day. Each period
first starts with:

 * a "stir" phase to dissipate heat and stir the water
 * a "normal" phase which does the main job of filtration
 * a "tank" phase which filters the tank

Filtering the tank in several shorter phases avoid the pool level to raise
too much. The gravity drain is unfortunately not able to sustain the pump
throughput even at the lowest level.