Regulate free chlorine
We allow 4 settings (low, mid, mid_high and high) for the free chlorine
setting and calculate the ORP setpoint based on the pH for that line. We
then regulate the ORP value using chlorine. Pretty much experimental, not
sure the P-controller will work properly. Needs a lot of testing.

Fixes #2