Add initial heating control
The heating will run once a day, at a given hour. It will run only if the
filtration is in eco mode. It will run until the setpoint is reached.

There is no "boost" feature implemented yet. However, changing the setpoint
will restart the heating period for the day if needed.

Fixes #1