Add enable switch for pH/ORP regulation
The regulation of pH and/or ORP can now be configured in the interface.
Measurements are still taken, only the feedback is forced to 0 if disabled.
Can be useful for testing or if you know you ran out of chemicals and don't
want to have the peristaltic pump running for nothing.

Fixes #11