Add periodic stir in eco_waiting
Removed the eco_stir state that was run at the beginning of a filtration
phase and implemented a periodic stir function that will run during
eco_waiting. This is to allow a more frequent stirring of the water since
a lot of heat builds up under the cover during hot days.