Postpone EEPROM writing on stop
We cannot write the position directly on a stop event because the motor
might still run for up to 3 seconds after we turn it off. We allow for
3 additional seconds for the cover to go outside of the end points and
save the position to EEPROM.

Also in case of OPEN, STOP, CLOSE events, nothing will be saved to EEPROM
if the transition is shorter than 3 seconds.