Add wintering mode to swim pump
If the temperature of the swim pump space is getting below zero, the pump
will switch on for one minute. There will be then a pause of at least 3
hours and if the temperature is again below zero, the pump will switch on
again.

Notice that it is still possible to turn off and manually control the pump
but it will not be possible to go back to the wintering mode. One has to
first stop the global mode and activate wintering again globally.