Add pump delay after heating
The heat pump requires the flow pump to run for at least 30 seconds after
it is done heating. A time.sleep(30) would do the trick but that would
block the actor and it would make Poupool look terribly wrong. Adding new
states to handle this makes the FSM more complicated but solves the issue
cleanly and render it more flexible in case things need to be adapted.

Notice that the case of the comfort mode has been solved differently in
e2149b14057