Dump dependencies to latest versions
Transitions 0.6.8 introduced a check to see if an attribute is already
present on an instance. Unfortunately this revealed that our "stop" state
is conflicting with the "stop" method from the actor. In fact, the model
we register is also a Pykka actor which has a stop method. The easiest fix
here is to just rename our "stop" state to something else. I chose "halt".