Change back signals to active HIGH
The relay board is active LOW but unfortunately, the relays get triggered
during the Arduino boot up. Looks like the signals on the pins are in a
kind of low state when the bootloader is running. The hardware has been
changed so that the relay are not active HIGH.

Notice that the pin states for the open/close events seem to be reversed.
I don't remember why I changed this in e900380e83 but I'll keep it like
that here.