Use threading.Timer instead of active polling
I initially wanted to avoid creating more threads but using a Timer here
is definitely much more efficient and simple. I think that "active" loop
was actually consuming most of the CPU. Also, the previous implementation
was not precise at all due to the fact that a delay was computed in each
iteration. The precision should be much better with the Timer now.