Upgrade to Pykka 2.0.0
Pykka 2.0.0 introduce the defer() method on the proxy. This is great for
fire and forget and it will now call the on_failure() callback in case
of exception. This is the behavior we want in most case in Poupool. So
now, if an actor raises an exception during a defer(ed) transition, an
exception will be printed in the log and Poupool will be terminated. We
don't need to rely on the DEBUG log level of the Pykka library to see the
problem.

Also we stop more actors when halt is triggered. A deadlock has been found
and workaround by using a timeout in the get() call.