Add settings for cover position in eco mode
This allows the pool to run in eco mode with the cover partially or fully
open. Especially useful for hot days (like now) to cool the pool at night
and still ensure chlorine and pH are ok.

It will never open the cover to match the setting. Only when closing.