Remove repeat decorator
If do_cancel() is called while the decorated method is still executing,
the cancel event will not be effective and the actor will continue running
repeating the calls. During state transitions, this was not really a
problem because often another delayed call would cancel the running one
anyway. But we have seen the problem with in some cases with the PWM and
this could be dangerous.